+++ plugins/modules/ping6.py
@@ -39,18 +39,21 @@ options:
         in the resource.
       - If no, the module will return error if any non-updatable parameters are provided.
     type: str
-  I:
+  interface:
     type: str
     description:
       - Network interface on which to ping, if you have multiple interfaces.
-  S:
+      - 'Deprecated alias: C(I).'
+  source_ip:
     type: str
     description:
       - Source IP address to be used in the outgoing query packets.
-  T:
+      - 'Deprecated alias: C(S).'
+  traffic_domain:
     type: int
     description:
       - Traffic Domain Id
+      - 'Deprecated alias: C(T).'
   V:
     type: int
     description:
@@ -69,10 +72,11 @@ options:
     type: str
     description:
       - Address of host to ping.
-  i:
+  interval:
     type: int
     description:
       - Waiting time, in seconds. The default value is 1 second.
+      - 'Deprecated alias: C(i).'
   m:
     type: bool
     description:
@@ -92,14 +96,16 @@ options:
     description:
       - Quiet output. Only summary is printed. For Nitro API, this flag is set by
         default
-  s:
+  packet_size:
     type: int
     description:
       - Data size, in bytes. The default value is 32.
-  t:
+      - 'Deprecated alias: C(s).'
+  timeout:
     type: int
     description:
       - Timeout in seconds before ping6 exits
+      - 'Deprecated alias: C(t).'
 extends_documentation_fragment: netscaler.adc.netscaler_adc
 
 """